requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
